+++ db/dbformat.h
@@ -250,7 +250,7 @@ class InternalKeyComparator
 
   virtual const char* Name() const override;
   virtual int Compare(const Slice& a, const Slice& b) const override;
-  virtual double Difference(const Slice&a, const Slice& b) const override;
+  virtual int64_t Difference(const Slice& a, const Slice& b) const override;
   // Same as Compare except that it excludes the value type from comparison
   virtual int CompareKeySeq(const Slice& a, const Slice& b) const;
   virtual void FindShortestSeparator(std::string* start,
@@ -705,8 +705,8 @@ struct RangeTombstone {
   }
 };
 
-inline double InternalKeyComparator::Difference(const Slice& akey,
-                                                const Slice& bkey) const {
+inline int64_t InternalKeyComparator::Difference(const Slice& akey,
+                                                 const Slice& bkey) const {
   // TODO: Provide a real implementation here.
   return 0;
 }
